pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
hello = "app.main:hello"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
tomli = "^2.0.1"
pytest = "^7.2.1"
pre-commit = "^3.0.0"
black = "^22.12.0"
flake8 = "^6.0.0"
isort = "^5.11.4"
interrogate = "^1.5.0"
flake8-pyproject = "^1.2.2"
recommonmark = "^0.7.1"
mkdocs-material = "^9.0.6"
mkdocstrings = {extras = ["python"], version = "^0.20.0"}
pygments = "^2.14.0"
click = "^8.1.3"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| build   
)/ 
'''

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503', 'F403', 'F401']
max-line-length = 79
max-complexity = 18
select = ['B','C','E','F','W','T4','B9']

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 95
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true
# generate-badge = "."
# badge-format = "svg"
